+++ CHANGELOG.rst
@@ -28,6 +28,7 @@ Fixed
 
 - **Projectroles**
     - Deprecated ``SODAR_API_*`` settings required in tests (#1495)
+    - Add workaround to ``ProjectInviteCreateView`` returning 404 with category and query string (#1510)
 
 
 v1.0.2 (2024-09-09)